refactor(framework): Move ServerApp to flwr.serverapp
