+++ framework/py/flwr/server/__init__.py
@@ -15,19 +15,38 @@
 """Flower server."""
 
 
+from typing import TYPE_CHECKING, Any
+
 from ..compat.server.app import start_server as start_server  # Deprecated
 from . import strategy
 from . import workflow as workflow
 from .client_manager import ClientManager as ClientManager
 from .client_manager import SimpleClientManager as SimpleClientManager
 from .compat import LegacyContext as LegacyContext
-from .grid import Driver as Driver
 from .grid import Grid as Grid
 from .history import History as History
 from .server import Server as Server
-from .server_app import ServerApp as ServerApp
 from .server_config import ServerConfig as ServerConfig
 from .serverapp_components import ServerAppComponents as ServerAppComponents
+
+if TYPE_CHECKING:
+    from flwr.compat.server.grid import Driver as Driver
+    from flwr.serverapp import ServerApp as ServerApp
+
+
+def __getattr__(name: str) -> Any:
+    """Lazily resolve compatibility exports."""
+    if name == "Driver":
+        from flwr.compat.server.grid import Driver  # pylint: disable=import-outside-toplevel
+
+        globals()[name] = Driver
+        return Driver
+    if name == "ServerApp":
+        from flwr.serverapp import ServerApp  # pylint: disable=import-outside-toplevel
+
+        globals()[name] = ServerApp
+        return ServerApp
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 __all__ = [
     "ClientManager",
