+++ framework/py/flwr/serverapp/__init__.py
@@ -16,9 +16,9 @@
 
 
 from flwr.server.grid import Grid
-from flwr.server.server_app import ServerApp
 
 from . import strategy
+from .server_app import ServerApp as ServerApp
 
 __all__ = [
     "Grid",
