+++ framework/py/flwr/simulation/run_simulation.py
@@ -42,7 +42,7 @@ from flwr.common.typing import Run
 from flwr.proto.task_pb2 import Task  # pylint: disable=E0611
 from flwr.server.grid import Grid, InMemoryGrid
 from flwr.server.run_serverapp import run as _run
-from flwr.server.server_app import ServerApp
+from flwr.serverapp import ServerApp
 from flwr.server.superlink.fleet import vce
 from flwr.server.superlink.fleet.vce.backend.backend import BackendConfig
 from flwr.server.superlink.linkstate import InMemoryLinkState, LinkStateFactory
